requirements.txt
flask
pandas
scikit-learn
groq
python-dotenv